Typo

+++ real_postprocessing/collect_latlon.py
@@ -144,7 +144,7 @@ def join_catalog_sel(search_dir, output_file, search_file = "", n = 0):
 					df.at[c, 'YR'], df.at[c, 'MO'], df.at[c, 'DY'] = data[1:4]			
 					df.at[c, 'HR'], df.at[c, 'MI'], df.at[c, 'SC'] = data[4].split(":")
 
-					if int(df.at[c, 'SC']) < 0:
+					if float(df.at[c, 'SC']) < 0:
 						ts = datetime.datetime.strptime("-".join(data[1:4]), "%Y-%m-%d-%H:%M")
 						ts -= datetime.timedelta(seconds = float(df[c, 'SC']))
 						df.at[c, 'timestamp'] = ts
@@ -168,7 +168,7 @@ def join_catalog_sel(search_dir, output_file, search_file = "", n = 0):
 
 					df.at[c, 'YR'], df.at[c, 'MO'], df.at[c, 'DY'] = data[0:3]			
 					df.at[c, 'HR'], df.at[c, 'MI'], df.at[c, 'SC'] = data[3:6]
-					if int(df.at[c, 'SC']) < 0:
+					if float(df.at[c, 'SC']) < 0:
 						ts = datetime.datetime.strptime("-".join(data[0:5]), "%Y-%m-%d-%H:%M")
 						ts -= datetime.timedelta(seconds = float(df[c, 'SC']))
 						df.at[c, 'timestamp'] = ts
